update deparment delete api

See merge request thebaapcompany/baap-academic-service!134

+++ routes/department.routes.js
@@ -6,14 +6,14 @@ const requestResponsehelper = require("@baapcompany/core-api/helpers/requestResp
 const ValidationHelper = require("@baapcompany/core-api/helpers/validation.helper");
 const TokenService = require("../services/token.services");
 const departmentModel = require("../schema/department.schema");
+const courseModel=require("../schema/courses.schema")
 const multer = require("multer");
 const upload = multer();
 const xlsx = require("xlsx");
 
 router.post(
     "/",
-    checkSchema(require("../dto/department.dto")),
-    //  TokenService.checkPermission(["EMD2"]),
+    checkSchema(require("../dto/department.dto")),  TokenService.checkPermission(["EMD2"]),
     async (req, res, next) => {
         if (ValidationHelper.requestValidationErrors(req, res)) {
             return;
@@ -35,22 +35,25 @@ router.get("/all", async (req, res) => {
     requestResponsehelper.sendResponse(res, serviceResponse);
 });
 
-router.delete("/groupId/:groupId/departmentId/:departmentId", TokenService.checkPermission(["EMD4"]), async (req, res) => {
+router.delete("/groupId/:groupId/departmentId/:departmentId",TokenService.checkPermission(["EMD4"]), async (req, res) => {
     try {
-
-        const groupId = req.params.groupId;
-        const departmentId = req.params.departmentId;
-        const Data = await service.deleteByDataId(groupId, departmentId);
-        if (!Data) {
-            res.status(404).json({ error: 'Department already exist for the provided course' });
-        } else {
-            res.status(201).json(Data);
-        }
+      const groupId = req.params.groupId;
+      const departmentId = req.params.departmentId;
+  
+      const hasAssignedCourses = await courseModel.exists({ groupId, departmentId });
+  
+      if (hasAssignedCourses) {
+        return res.status(401).json({ error: 'Department has assigned courses ' });
+      }
+  
+      const data = await service.deleteByDataId(groupId, departmentId);
+      res.status(201).json(data);
     } catch (error) {
-        console.error(error);
-        res.status(500).json({ error: 'Internal Server Error' });
+      console.error(error);
+      res.status(500).json({ error: 'Internal Server Error' });
     }
-});
+  });
+  
 
 
 router.get("/all/getByGroupId/:groupId", TokenService.checkPermission(["EMD1"]),
